Implemented advanced searches, fixed validations, exceptions and code failures

+++ src/main/java/com/library/api/service/BookService.java
@@ -3,13 +3,19 @@ package com.library.api.service;
 import com.library.api.dto.BookDTO;
 import com.library.api.entities.Book;
 import com.library.api.repository.BookRepository;
+import com.library.api.service.exceptions.DatabaseException;
+import com.library.api.service.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Instant;
 import java.time.LocalDate;
+import java.time.ZoneId;
 
 @Service
 public class BookService {
@@ -27,10 +33,14 @@ public class BookService {
 
     @Transactional
     public BookDTO updateBook(Long id, BookDTO bookDto) {
-        Book book = repository.getReferenceById(id);
-        copyDtoToBook(bookDto, book);
-        book = repository.save(book);
-        return new BookDTO(book);
+        try {
+            Book book = repository.getReferenceById(id);
+            copyDtoToBook(bookDto, book);
+            book = repository.save(book);
+            return new BookDTO(book);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Book does not exist");
+        }
     }
 
     @Transactional(readOnly = true)
@@ -41,19 +51,35 @@ public class BookService {
 
     @Transactional(readOnly = true)
     public BookDTO searchBookById(Long id) {
-        Book book = repository.getReferenceById(id);
+        Book book = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Book does not exist"));
         return new BookDTO(book);
+    }
+
+    @Transactional(readOnly = true)
+    public Page<BookDTO> searchBookByAuthor(String author, Pageable pageable) {
+        Page<BookDTO> book = repository.searchBookByAuthor(author, pageable);
+        return book.map(BookDTO::new);
     }
 
     @Transactional
     public void deleteBookById(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Book does not exist");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Referential integrity failure");
+        }
     }
 
     public void copyDtoToBook(BookDTO bookDto, Book book) {
 
-        if (!bookDto.getName().isEmpty()) {
-            book.setName(bookDto.getName());
+        LocalDate today = LocalDate.ofInstant(Instant.now(), ZoneId.systemDefault());
+        LocalDate yearOfPublication = LocalDate.parse(bookDto.getYearOfPublication());
+
+        if (!bookDto.getTitle().isEmpty()) {
+            book.setTitle(bookDto.getTitle());
         }
 
         if (!bookDto.getAuthor().isEmpty()) {
@@ -61,7 +87,14 @@ public class BookService {
         }
 
         if (!bookDto.getYearOfPublication().isEmpty()) {
+            if (yearOfPublication.isAfter(today)) {
+                throw new ResourceNotFoundException("Date cannot be greater than the current date");
+            }
             book.setYearOfPublication(LocalDate.parse(bookDto.getYearOfPublication()));
+        }
+
+        if (bookDto.getPriceDayRent() != 0) {
+            book.setPriceDayRent(bookDto.getPriceDayRent());
         }
     }
 }
